Enum wasmer_derive::parse::WasmerAttr

pub enum WasmerAttr {
    Export {
        identifier: Option<LitStr>,
        optional: bool,
        aliases: Vec<LitStr>,
        span: Span,
    },
}

Variants

Export

Fields

identifier: Option<LitStr>

The identifier is an override, otherwise we use the field name as the name to lookup in instance.exports.

optional: bool

aliases: Vec<LitStr>

span: Span